Create disease_drugseries_TF-IDF_weighted_LSA_DSS_cosinesim
Version of the core Python code block that scrapes chronologically designated NLM Pubmed abstracts related to a disease and any desired co-terms (the query) and then provides quantitative document similarity data (cosine similarities) relative to documents separately scraped from a series of drugs of interest, with the document comparison enhanced by TF-IDF weighting and latent semantic analysis (LSA), which implements topic modeling to discern similarities among Deep Semantic Structures (DSS's) beyond direct term matches alone. For example, if you are interested in von Willebrand disease and want to check for overlap of abstracts pertaining to VWD and digoxin, vasopressin (positive control), furosemide, menaquinone, diltiazem, and enalapril for the years 2005-2007 will return a list of cosine similarities showing which drug abstracts had the greatest overlap -- in both direct term matches and deep semantic structures -- with VWD-related abstracts during that period (with optional co-terms helping to narrow the search for investigations of overlap with drug mechanism or pharmacokinetics, for example). This code block can be universally extended to test for such document comparisons across any range of compared topics simply by altering the input queries. The training of the LSA model in this case is accomplished using neutral pertinent corpora (semantic spaces) loaded locally from disk, consisting of abstracts without a specific topic focus scraped across a given chronological range (see the CorpusOptima repository for details on how the semantic spaces have been assembled), though any appropriate semantic space can be used for the LSA training as needed. The code her has been carefully documented to lay out the basis for each of the steps and, in particular, for the often intricate text editing steps as well as the tokenization, vectorization, and training of the gensim TF-IDF and LSA models. Dependencies outside the standard Python library include BeautifulSoup and Gensim. Note that for Gensim to function most effectively, a version of numpy with the Intel MKL (math kernel library) as well as a package to assist with rapid linear algebra calculations and matrix handling (such as OpenBLAS) should be installed, though most of the additional modules can be found in a standard distribution such as the Anaconda distribution.